Refactored Homepage code to make it more legible
The goal is to now put each "section" in its own component,
to help with legibility and code maintenance.

TODO
- Finish Other Sections
- Finish Home CSS
- Eventually the About Page.
